requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.3.1
onnxruntime>=1.8.1
numpy>=1.21.2
python-dotenv>=0.19.1
pydantic>=1.8.2
jinja2>=3.0.1
pytest>=6.2.5
httpx>=0.19.0